moved useeffect hooks to respective components for more modular code
